requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
DateTime==5.5
Django==5.1.4
django-cors-headers==4.6.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
pillow==11.0.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
pytz==2024.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
zope.interface==7.2
